Delete ArticleState.kt
