+++ app/src/main/java/ru/skillbranch/skillarticles/data/repositories/ArticleRepository.kt
@@ -1,11 +1,13 @@
 package ru.skillbranch.skillarticles.data.repositories
 
 import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
 import ru.skillbranch.skillarticles.data.*
 
 object ArticleRepository {
     private val local = LocalDataHolder
     private val network = NetworkDataHolder
+    private val isSearchLiveData = MutableLiveData<Boolean>()
 
     fun loadArticleContent(articleId: String): LiveData<List<Any>?> {
         return network.loadArticleContent(articleId) //5s delay from network
@@ -25,5 +27,11 @@ object ArticleRepository {
 
     fun updateArticlePersonalInfo(info: ArticlePersonalInfo) {
         local.updateArticlePersonalInfo(info)
+    }
+
+    fun getSearchStatus() = isSearchLiveData
+
+    fun updateSearchStatus(status:Boolean){
+        isSearchLiveData.value = status
     }
 }
