--- app/src/main/java/ru/skillbranch/skillarticles/viewmodels/ArticleState.kt
@@ -1,25 +1,0 @@
-package ru.skillbranch.skillarticles.viewmodels
-
-data class ArticleState(
-    val isAuth:Boolean = false, //пользователь авторизован
-    val isLoadingContent: Boolean = true, //контент загружается
-    val isLoadingReviews:Boolean = true, //отзывы загружаются
-    val isLike:Boolean = false, //отмечено как Like
-    val isBookmark:Boolean = false, //в закладках
-    val isShowMenu:Boolean = false, //отображается меню
-    val isBigText: Boolean = false, //шрифт увеличен
-    val isDarkMode:Boolean = false, //темный режим
-    val isSearch:Boolean = false, //режим поиска
-    val searchQuery:String? = null, //поисковый запрос
-    val searchResults:List<Pair<Int, Int>> = emptyList(), //результаты поиска (стартовая и конечная позиции)
-    val searchPosition: Int = 0, //текущая позиция найденного результата
-    val shareLink:String? = null, //ссылка Share
-    val title: String? = null, //заголовок статьи
-    val category: String? = null, //категория
-    val categoryIcon:Any? = null, //иконка категории
-    val date:String? = null, //дата публикации
-    val author: Any? = null, //автор статьи
-    val poster:String? = null, //обложка статьи
-    val content: List<Any> = emptyList(), //контент
-    val reviews:List<Any> = emptyList() //комментарии
-)
